Rename browser save core module
